requirements.txt
boto3
elasticsearch
fire
flask-appbuilder
flask-bootstrap
fsspec
pandas
pandasql
# pandas-profiling
psycopg2-binary
s3fs
sqlparse
